+++ backend/app/models/code.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class Code(BaseModel):
+    language: str
+    prerequisites: Optional[str] = ""
+    code: str
