--- backend/constants.py
@@ -1,7 +1,0 @@
-ACTION_JOIN = "join"
-ACTION_JOINED = "joined"
-ACTION_DISCONNECTED = "disconnected"
-ACTION_LANG_CHANGE = "lang_change"
-ACTION_CODE_CHANGE = "code_change"
-ACTION_SYNC_CODE = "sync_code"
-ACTION_LEAVE = "leave"
